02-wordpress/02-products/requirements.txt
Pillow
pyairtable
python-dotenv
requests
woocommerce